-fixes backported from Kexi 0.1 -a few more unnecessary mesages are now hidden from translation
svn path=/trunk/koffice/; revision=408569

+++ kexi/plugins/forms/kexiformscrollview.h
@@ -84,7 +84,7 @@ class KEXIFORMUTILS_EXPORT KexiFormScrollView :
 			if (!item)
 				return -1;
 			KexiFormDataItemInterfaceToIntMap::ConstIterator it(m_fieldNumbersForDataItems.find( item ));
-			return it!=m_fieldNumbersForDataItems.constEnd() ? it.data() : -1;
+			return it!=m_fieldNumbersForDataItems.constEnd() ? (int)it.data() : -1;
 		}
 
 		/*! @internal Used by KexiFormView in view switching. */
